# swagger-openapi.yaml
openapi: 3.0.1
info:
  title: "API Avaliação Sprint 4"
  description: "API para listar pessoas, produtos e pedidos de um sistema."
  termsOfService: "http://swagger.io/terms/"
  contact:
    name: "Alexys Santiago"
    email: "alexys.fernande_BOLS@compasso.com.br"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0"
  version: "1.0"
servers:
- url: "http://localhost:8080"
  description: "API do servidor local"
...
paths:
...
  /protected/pedido:
    get:
      tags:
      - PedidoController
      summary: "listar"
      parameters:
      - name: "Authorization"
        in: header
        description: "Header para token JWT"
        schema:
          type: string
      responses:
        200:
          description: "OK"
          content:
            "*/*":
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PedidoDto"
        403:
          description: "Forbidden"
          content:
            application/json:
              example:
                - "timestamp": "2021-07-18T00:04:26.475+00:00"
                  "status": 403
                  "error": "Forbidden"
                  "message": "Access Denied"
                  "path": "/protected/pedido"
    post:
      tags:
      - PedidoController
      summary: "cadastrar"
      parameters:
      - name: "Authorization"
        in: header
        description: "Header para token JWT"
        schema:
          type: string
      requestBody:
        description: "pedidoForm"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoForm"
        required: true
      responses:
        201:
          description: "Created"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PedidoDto"
        400:
          description: "Bad Request"
          content: {}
        403:
          description: "Forbidden"
          content:
            application/json:
              example:
                - "timestamp": "2021-07-18T00:04:26.475+00:00"
                  "status": 403
                  "error": "Forbidden"
                  "message": "Access Denied"
                  "path": "/pessoa/01898745028"
  /protected/pedido/{id}:
    get:
      tags:
      - PedidoController
      summary: "detalhar"
      parameters:
      - name: "Authorization"
        in: header
        description: "Header para token JWT"
        schema:
          type: string
      - name: id
        in: path
        description: "id do pedido"
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: "OK"
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PedidoDto"
        403:
          description: "Forbidden"
          content:
            application/json:
              example:
                - "timestamp": "2021-07-18T00:04:26.475+00:00"
                  "status": 403
                  "error": "Forbidden"
                  "message": "Access Denied"
                  "path": "/protected/pedido"
        404:
          description: "Not Found"
          content: {}
    put:
      tags:
      - PedidoController
      summary: "atualizar"
      parameters:
      - name: "Authorization"
        in: header
        description: "Header para token JWT"
        schema:
          type: string
      - name: id
        in: path
        description: "id do pedido"
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: "pedidoForm"
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PedidoForm"
        required: true
      responses:
        200:
          description: "OK"
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/PedidoDto"
        403:
          description: "Forbidden"
          content:
            application/json:
              example:
                - "timestamp": "2021-07-18T00:04:26.475+00:00"
                  "status": 403
                  "error": "Forbidden"
                  "message": "Access Denied"
                  "path": "/protected/pedido"
        404:
          description: "Not Found"
          content: {}
    delete:
      tags:
      - PedidoController
      summary: "remover"
      parameters:
      - name: "Authorization"
        in: header
        description: "Header para token JWT"
        schema:
          type: string
      - name: id
        in: path
        description: "id do pedido"
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: "OK"
          content:
            "*/*":
              schema:
                type: object
        403:
          description: "Forbidden"
          content:
            application/json:
              example:
                - "timestamp": "2021-07-18T00:04:26.475+00:00"
                  "status": 403
                  "error": "Forbidden"
                  "message": "Access Denied"
                  "path": "/protected/pedido"
        404:
          description: "Not Found"
          content: {}
...
components:
  schemas:
...
    PedidoDto:
      title: "PedidoDto"
      type: object
      properties:
        total:
          type: number
        dataCriacao:
          type: string
        produtoDto:
          type: array
          items:
            $ref: "#/components/schemas/ProdutoDto"
      example:
        total: 22.5
        dataCriacao: "2021-07-17 14:03:47"
        produtoDto:
          - id: 1
            descricao: "Arroz"
            precoUnitario: 10
          - id: 2
            descricao: "Feijão"
            precoUnitario: 5
          - id: 3
            descricao: "Açúcar"
            precoUnitario: 3.5
          - id: 4
            descricao: "Lentilha"
            precoUnitario: 4
    PedidoForm:
      title: "PedidoForm"
      type: object
      properties:
        idProduto:
          type: array
          items:
            type: integer
            format: int64
      example:
        idProduto:
          - 1
          - 2
          - 3
          - 4
...
    ProdutoDto:
      title: "ProdutoDto"
      type: object
      properties:
        id:
          type: integer
          format: int64
        descricao:
          type: string
        precoUnitario:
          type: number
      example:
        id: 7
        descricao: "Coxinha"
        precoUnitario: 3.99
...